Use medium precision for quad colors

+++ src/renderer/quads/shaders/fragment.glsl
@@ -1,4 +1,4 @@
-varying highp float v_color;
+varying mediump float v_color;
 
 void main(void) {
     gl_FragColor = vec4(v_color, v_color, v_color, 1);
